==> Part 1 - Data Preparation.ipynb <==
# Project 3, Part 1 - Loading/Cleaning Data
* Jacob Wang
* 1/11/2023
* Coding Dojo Data Science Project 3

---------------------------------------------------------

For this project, you have been hired to produce a MySQL database on Movies from a subset of IMDB's publicly available dataset. Ultimately, you will use this database to analyze what makes a movie successful and will provide recommendations to the stakeholder on how to make a successful movie.
_________________________________________________________

## Specifications
Your stakeholder only wants you to include information for movies based on the following specifications:

* Exclude any movie with missing values for genre or runtime
* Include only full-length movies (titleType = "movie").
* Include only fictional movies (not from documentary genre)
* Include only movies that were released 2000 - 2021 (include 2000 and 2021)
* Include only movies that were released in the United States

import pandas as pd
import seaborn as sns
import numpy as np

# basics

basics_url = 'https://datasets.imdbws.com/title.basics.tsv.gz'
akas_url = 'https://datasets.imdbws.com/title.akas.tsv.gz'
ratings_url = 'https://datasets.imdbws.com/title.ratings.tsv.gz'

basics = pd.read_csv(basics_url, sep='\t', low_memory=False)

ratings = pd.read_csv(ratings_url, sep='\t', low_memory=False)

basics.info()

akas = pd.read_csv(akas_url, sep='\t', low_memory=False)
akas.head()

# convert nulls
basics = basics.replace({'\\N':np.nan})

# eliminate null runtimeMinutes
basics.dropna(subset=['runtimeMinutes'], inplace=True)
basics.isna().sum()

# eliminate nulls in 'genres'

basics.dropna(subset=['genres'], inplace=True)
basics.isna().sum()

# keep only titleType == 'movie'
basics['titleType'].value_counts()
movie_filter = basics['titleType'] == 'movie'
basics.loc[movie_filter, :].head()

# keep 'startYear' between 2000-2022
yrFilter = basics['startYear'] >= '2000'
yrFilter2 = basics['startYear'] < '2022'

basics = basics.loc[movie_filter]
basics = basics.loc[yrFilter & yrFilter2]
basics.info()
basics.head()

basics = basics[(basics['genres'].notna())]

# eliminate documentaries 
# Exclude movies that are included in the documentary category.
is_documentary = basics['genres'].str.contains('documentary',case=False)
basics = basics[~is_documentary]

basics.info()
basics.head()


# akas


akas.info()
akas.head()

# convert nulls
akas = akas.replace({'\\N':np.nan})

us_filter = akas['region'] == 'US'
akas = akas[(us_filter)] 

# Filter the basics table down to only include the US by using the filter akas dataframe
keepers =basics['tconst'].isin(akas['titleId'])
keepers

# filter basics
basics = basics[keepers]
basics.head()

# Ratings

ratings.info()
ratings.head()

# convert nulls
ratings = ratings.replace({'\\N':np.nan})

# Filter the ratings table down to only include the US by using the filter akas dataframe
keepers = ratings['tconst'].isin(akas['titleId'])
ratings = ratings[keepers]
ratings.info()
ratings.head()

# save DFs to Data folder

## Save all dataframes to file.

akas.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)
ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)

# save basics df
basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)

# Open saved file and preview again
df = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
df.info()
df.head()

# Open saved file and preview again
df2 = pd.read_csv("Data/title_akas.csv.gz", low_memory = False)
df2.info()
df2.head()

# Open saved file and preview again
df3 = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
df3.info()
df3.head()
