==> Part 2 - API Data Extraction.ipynb <==
# Project 3, Part 2.1 - Efficient TMDB API Calls

* Jacob Wang
* Coding Dojo Data Science Project 3
* 1/11/2023


Your Stakeholder Wants More Data!

After investigating the preview of your data from Part 1, your stakeholder realized that there is no financial information included in the IMDB data (e.g. budget or revenue).

This will be a major roadblock when attempting to analyze which movies are successful and must be addressed before you will be able to determine which movies are successful. Your stakeholder identified The Movie Database (TMDB) as a great source of financial data (https://www.themoviedb.org/). Thankfully, TMDB offers a free API for programmatic access to their data!

Your stakeholder wants you to extract the budget, revenue, and MPAA Rating (G/PG/PG-13/R), which is also called "Certification".

## Specifications - Financial Data
* Your stakeholder would like you to extract and save the results for movies that meet all of the criteria established in part 1 of the project (You should already have a filtered dataframe saved from part one as a csv.gz file)

* As a proof-of-concept, they requested you perform a test extraction of movies that started in 2000 or 2001

* Each year should be saved as a separate .csv.gz file

* One function will add the certification (MPGG Rating) to movie.info

* The other function will help you append/extend a JSON file with Python

## Confirm Your API Function works.
In order to ensure your function for extracting movie data from TMDB is working, test your function on these 2 movie ids:

* tt0848228 ("The Avengers")
* tt0332280 ("The Notebook").

Make sure that your function runs without error and that it returns the correct movie's data for both test ids.

# imports
import json
import pandas as pd
import tmdbsimple as tmdb
import os, time
from tqdm.notebook import tqdm_notebook

# API Key

with open('/Users/jacobwang/.secret/tmdb_api.json') as f: #change the path to match YOUR path!!
    login = json.load(f)
login.keys()

tmdb.API_KEY =  login['api-key']

FOLDER = "Data/"
os.makedirs(FOLDER, exist_ok=True)
os.listdir(FOLDER)

## Define functions:

def write_json(new_data, filename): 
    """Appends a list of records (new_data) to a json file (filename). 
    Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/"""  
    
    with open(filename,'r+') as file:
        # First we load existing data into a dict.
        file_data = json.load(file)
        ## Choose extend or append
        if (type(new_data) == list) & (type(file_data) == list):
            file_data.extend(new_data)
        else:
             file_data.append(new_data)
        # Sets file's current position at offset.
        file.seek(0)
        # convert back to json.
        json.dump(file_data, file)

def get_movie_with_rating(movie_id):
    movie = tmdb.Movies(movie_id)
    
    info = movie.info()
    
    releases = movie.releases()
    
    for c in releases['countries']:
            if c['iso_3166_1'] == 'US':
                info['certification'] = c['certification']
    return info

# Load in the dataframe from project part 1 as basics:
basics = pd.read_csv('Data/title_basics.csv.gz')
basics.head()

# Define a list of the Years to Extract from the API
YEARS_TO_GET = [2000, 2001]

# define 'errors' list
errors = [ ]

## Set up progress bar 

def movie_year_ratings(YEARS_TO_GET):
# Start of OUTER loop
    for YEAR in tqdm_notebook(YEARS_TO_GET, desc='YEARS', position=0):
    #Defining the JSON file to store results for year
        JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    # Check if file exists
        file_exists = os.path.isfile(JSON_FILE)
    # If it does not exist: create it
        if file_exists == False:
# save an empty dict with just "imdb_id" to the new json file.
            with open(JSON_FILE,'w') as f:
                json.dump([{'imdb_id':0}],f)
            #Saving new year as the current df
        df = basics.loc[ basics['startYear']==YEAR].copy()
            # saving movie ids to list
        movie_ids = df['tconst'].copy()
        # Load existing data from json into a dataframe called "previous_df"
        previous_df = pd.read_json(JSON_FILE)
            # filter out any ids that are already in the JSON_FILE
        movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]
                #Get index and movie id from list
    # INNER Loop
        for movie_id in tqdm_notebook(movie_ids_to_get,
                                  desc=f'Movies from {YEAR}',
                                  position=1,
                                  leave=True):
            try:
            # Retrieve then data for the movie id
                temp = get_movie_with_rating(movie_id)  
            # Append/extend results to existing file using a pre-made function
                write_json(temp,JSON_FILE)
            # Short 20 ms sleep to prevent overwhelming server
                time.sleep(0.02)
            
            except Exception as e:
                errors.append([movie_id, e])
                
    # save results as csv.gz file
        final_year_df = pd.read_json(JSON_FILE)
        final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", compression="gzip", index=False)
    print(f"- Total errors: {len(errors)}")

# test function
movie_year_ratings(YEARS_TO_GET)

## Verify API Function 

# extract movie data for Avengers/Notebook
## testing our function by looping through a list of ids
import pandas as pd
test_ids = ["tt0848228","tt0332280"]
results = []
for movie_id in test_ids:
    
    try:
        movie_info = get_movie_with_rating(movie_id)
        results.append(movie_info)
        
    except: 
        pass
    
pd.DataFrame(results)


movie_info
