==> Part 3 - Database Creation (MySQL).ipynb <==
# Project 3, Part 3 - MySQL Database 
* Jacob Wang
* 1/11/2023
* Coding Dojo Data Science Project 3 

import json
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

import pymysql
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists
pymysql.install_as_MySQLdb()

basics = pd.read_csv('Data/title_basics.csv.gz')
basics.info()
basics.head()

## create a col with a list of genres
basics['genres_split'] = basics['genres'].str.split(',')
basics

# separate the list of genres into new rows
exploded_genres = basics.explode('genres_split')
exploded_genres

# get unique genres
unique_genres = sorted(exploded_genres['genres_split'].unique())

## Create title_genres table


title_genres = exploded_genres[['tconst', 'genres_split']].copy()
title_genres.head()

## Create a genre mapper dictionary to replace string genres with integers

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

## Make dictionary with list of unique genres as key, new int id as values
genre_id_map = dict(zip(unique_genres, range(len(unique_genres))))
genre_id_map

## Replace the string genres in title_genres with the new integer ids

## make new integer genre_id and drop string genres
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres = title_genres.drop(columns = 'genres_split')
title_genres.head()

## Convert the genre map dictionary into a dataframe

# manually make df with named cols from .keys and .values
genre_lookup = pd.DataFrame({'genre_name': genre_map.keys(),
                            'genre_id': genre_map.values()})
genre_lookup.head()

# Save MySQL tables with tconst as the primary key

basics.dtypes

## create connection
with open('/Users/jacobwang/.secret/mysql.json') as f:
    login = json.load(f)
login.keys()

connection = f"mysql+pymysql://{login['username']}:{login['password']}@localhost/movies"
engine = create_engine(connection)

# Check if the database exists. If not, create it.
if database_exists(connection) == False:
  create_database(connection)
else:
  print('The database already exists')

basics.head()


basics = basics.drop(columns = ['originalTitle', 'isAdult', 'titleType', 'genres','genres_split'])

## Example
from sqlalchemy.types import *
## Calculate max string lengths for object columns
key_len = basics['tconst'].fillna('').map(len).max()
title_len = basics['primaryTitle'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Float(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}

# Save to sql with dtype and index=False
basics.to_sql('title_basics',engine,dtype=df_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

title_genres.to_sql('title_genres',engine, if_exists='replace',index=False)

genre_lookup.to_sql('genre_lookup',engine, if_exists='replace',index=False)

engine.execute('ALTER TABLE genre_lookup ADD PRIMARY KEY (`genre_id`);')

# run queries to test database
q = """
SELECT * FROM title_basics
LIMIT 5;"""
pd.read_sql(q, engine)

# run queries to test database
q = """
SELECT * FROM title_genres
LIMIT 5;"""
pd.read_sql(q, engine)

q = """
SELECT * FROM genre_lookup
LIMIT 5;"""
pd.read_sql(q, engine)

## Create 'ratings' table

ratings = pd.read_csv('Data/title_ratings.csv.gz')

ratings.info()
ratings.head()

key_len = ratings['tconst'].fillna('').map(len).max()
# create schema for ratings
ratings_schema = {
    "tconst": String(key_len+1), 
    'averageRating':Float(),
    'numVotes':Integer()}

ratings.to_sql('title_ratings',engine,dtype=ratings_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

q = """
SELECT * FROM title_ratings 
LIMIT 5; """
pd.read_sql(q, engine)

## Create tables for tmdb data

tmdb2000 = pd.read_csv('Data/final_tmdb_data_2000.csv.gz')
tmdb2001 = pd.read_csv('Data/final_tmdb_data_2001.csv.gz')
tmdb_data = pd.concat([tmdb2000, tmdb2001], axis=0)

tmdb_data.info() 
tmdb_data.head()

# drop unnecessary columns
tmdb_data = tmdb_data[['imdb_id', 'budget', 'revenue', 'certification']]
tmdb_data

## Calculate max string lengths for object columns
key_len = tmdb_data['imdb_id'].str.len().max()
cert_len = tmdb_data['certification'].str.len().max()
## Create a schema dictonary using Sqlalchemy datatype objects
tmdb_schema = {"imdb_id": String(key_len+1), 
    "certification": Text(cert_len+1),
    'budget':Float(),
    'revenue':Float()}

tmdb_data.to_sql('tmdb_data',engine, dtype=tmdb_schema, if_exists='replace',index=False)

engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

q = """
SELECT * FROM tmdb_data
LIMIT 5;"""
pd.read_sql(q, engine)

## Confirm all tables have been created

q = """SHOW TABLES"""
pd.read_sql(q, engine)
